// src/modules/stripe-payment/utils/stripe-payment.enum.ts
export enum UI_MODE {
  EMBEDDED = "embedded",
}

export enum PAYMENT_MODE {
  PAYMENT = "payment",
}

export enum CURRENCY {
  USD = "usd",
}

export enum REDIRECT_ON_COMPLETION {
  NEVER = "never",
}

export enum DURATION {
  MONTHLY = "monthly",
  YEARLY = "yearly",
  WEEKLY = "weekly",
}

export enum PLATFORM_FEE_PERCENTAGE {
  PERCENTAGE = 0.1, // 10%
}